Break out the wait_screen_ui
