+++ crates/re_viewer/src/app.rs
@@ -705,7 +705,7 @@ impl eframe::App for App {
                         render_ctx.begin_frame();
 
                         if store_db.is_empty() {
-                            wait_screen_ui(ui, &self.rx);
+                            crate::ui::wait_screen_ui(ui, &self.rx);
                         } else {
                             self.state.show(
                                 &mut blueprint,
@@ -722,7 +722,7 @@ impl eframe::App for App {
                         }
                     }
                 } else {
-                    wait_screen_ui(ui, &self.rx);
+                    crate::ui::wait_screen_ui(ui, &self.rx);
                 }
             });
 
@@ -794,60 +794,6 @@ fn paint_native_window_frame(egui_ctx: &egui::Context) {
         re_ui::ReUi::native_window_rounding(),
         stroke,
     );
-}
-
-fn wait_screen_ui(ui: &mut egui::Ui, rx: &Receiver<LogMsg>) {
-    ui.centered_and_justified(|ui| {
-        fn ready_and_waiting(ui: &mut egui::Ui, txt: &str) {
-            let style = ui.style();
-            let mut layout_job = egui::text::LayoutJob::default();
-            layout_job.append(
-                "Ready",
-                0.0,
-                egui::TextFormat::simple(
-                    egui::TextStyle::Heading.resolve(style),
-                    style.visuals.strong_text_color(),
-                ),
-            );
-            layout_job.append(
-                &format!("\n\n{txt}"),
-                0.0,
-                egui::TextFormat::simple(
-                    egui::TextStyle::Body.resolve(style),
-                    style.visuals.text_color(),
-                ),
-            );
-            layout_job.halign = egui::Align::Center;
-            ui.label(layout_job);
-        }
-
-        match rx.source() {
-            re_smart_channel::SmartChannelSource::Files { paths } => {
-                ui.strong(format!(
-                    "Loading {}…",
-                    paths
-                        .iter()
-                        .format_with(", ", |path, f| f(&format_args!("{}", path.display())))
-                ));
-            }
-            re_smart_channel::SmartChannelSource::RrdHttpStream { url } => {
-                ui.strong(format!("Loading {url}…"));
-            }
-            re_smart_channel::SmartChannelSource::RrdWebEventListener => {
-                ready_and_waiting(ui, "Waiting for logging data…");
-            }
-            re_smart_channel::SmartChannelSource::Sdk => {
-                ready_and_waiting(ui, "Waiting for logging data from SDK");
-            }
-            re_smart_channel::SmartChannelSource::WsClient { ws_server_url } => {
-                // TODO(emilk): it would be even better to know whether or not we are connected, or are attempting to connect
-                ready_and_waiting(ui, &format!("Waiting for data from {ws_server_url}"));
-            }
-            re_smart_channel::SmartChannelSource::TcpServer { port } => {
-                ready_and_waiting(ui, &format!("Listening on port {port}"));
-            }
-        };
-    });
 }
 
 impl App {
