+++ crates/re_viewer/src/ui/mod.rs
@@ -4,6 +4,7 @@ mod blueprint_sync;
 mod rerun_menu;
 mod selection_history_ui;
 mod top_panel;
+mod wait_screen_ui;
 
 pub(crate) mod memory_panel;
 pub(crate) mod selection_panel;
@@ -12,4 +13,5 @@ pub(crate) mod selection_panel;
 
 pub(crate) use {
     self::blueprint::Blueprint, self::rerun_menu::rerun_menu_button_ui, self::top_panel::top_panel,
+    self::wait_screen_ui::wait_screen_ui,
 };
